Add support for Google’s Location Services API (#185)
* feat(android-backend): refactor module

* feat(android): play services manager

* feat(android): fix permission crash
